Removed the --(no)parallel option.
After doing internal tests to optimize the parallelization of BLAST and HMMER, we found that BLAST tends to be faster in multithread option than in parallel one. However, HMMER had the opposite behavior. Then, to speed up the program, BLAST will be launched using multiple threads and HMMER in parallel.