Update virannot.py
v 0.10.0: * Added the prediction of the origin and terminus of replication for circular contigs based on the cumulative GC skew. After detecting the origin coordinate, the chromosome is realigned from the origin. As a consequence of that, two new parameters ("--windowsize" and "--slidingsize") were added to determine the window size and the sliding window size respectively.
               * Fixed error with the start position of the genes in the GenBank files, which were not related to the amino acid sequences and made that the sequence length was not multiple of three.
               * Added "/locus_tag" in the putative genes in the GenBank files